telegramBot/requirements.txt
functions-framework==3.*
requests==2.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
